+++ src/constants/ExtensionState.ts
@@ -3,5 +3,6 @@ export const ExtensionState = {
   PagesView: `frontMatter:Pages:ViewType`,
   SelectedFolder: `frontMatter:SelectedFolder`,
   Version: `frontMatter:Version`,
-  SettingPromoted: `frontMatter:Settings:Promoted`
+  SettingPromoted: `frontMatter:Settings:Promoted`,
+  MoveTemplatesFolder: `frontMatter:Templates:Move`,
 };