+++ src/constants/index.ts
@@ -3,6 +3,7 @@ export * from './DefaultFields';
 export * from './Extension';
 export * from './ExtensionState';
 export * from './Links';
+export * from './LocalStore';
 export * from './Navigation';
 export * from './charMap';
 export * from './context';